Create NumberGuessingGame.java
This is a number guessing game in Java, where the user has to guess a randomly generated number within 10 attempts. The program provides hints to the user to guide them towards the correct number. If the user is unable to guess the correct number within the specified number of attempts, the game ends.